WIP: change how emulator works
 * install platform-tools inside emulator
 * continue to start an adb server inside the emulator container
 * allow connecting to the emulator from an adb server outside the
   emulator container
 * use ANDROID_SERIAL to select an emulator rather than ADB_SERVER_SOCKET

Goals:

 * Make it easier to start multiple emulators concurrently and run tests
   on them.
 * Make it easier to run Android libc++ tests from outside the buildbot
   Docker container.
 * Make it easier to just run a subset of the tests, rather than invoking
   run-buildbot to run them all in one go.
    - see libcxx/utils/libcxx-lit for an example of the ordinary workflow

TODO: What about adb protocol version mismatches? (Oh wait, is that fixed
by switching to ANDROID_SERIAL from ADB_SERVER_SOCKET? i.e. by ensuring
that the adb client and server are always the same binary? I think that's
what I did that? It's been a while. :) )

In principle, we could version-lock platform-tools between the buildbot
and emulator Docker images by upgrading them together, but that's not a
guarantee. It's harder to do that between the emulator Docker image and
a local dev environment. The current strategy of copying platform-tools
into the emulator seems problematic for a variety of reasons:

 * Where do we copy *from* ($ANDROID_SDK/platform-tools? But that env var
   might not be defined). It's common for Android devs to have multiple
   copies of adb on their machines. A copy of adb isn't necessarily part
   of a platform-tools package.

 * Actually getting adb into the Docker container is a pain. The current
   approach is to mount an extra, special, volume and copy it to/from at
   Docker container startup. It's cumbersome.

Better to just use adb in a more normal way and not worry about restarting
the adb server between test runs. (At least, not worry as much. If it
actually becomes a problem, then we can add code to try and clean up the
adb server, in the CI scripts.)